+++ package.json
@@ -9,7 +9,8 @@
     "install": "./install.sh",
     "build": "./build.sh",
     "server": "[ -f server.js ] && node server || ts-node src/server",
-    "server:dist": "cd dist && node server"
+    "server:dist": "cd dist && node server",
+    "deploy_algorand": "cd ./src/web3/algorand && ts-node deploy.ts"
   },
   "author": "",
   "license": "MIT",
@@ -17,22 +18,29 @@
     "@influxdata/influxdb-client": "^1.31.0",
     "@onflow/fcl": "^1.2.0",
     "@sentry/node": "^6.17.8",
+    "algosdk": "^1.19.0",
     "axios": "^0.26.1",
     "base58-js": "^1.0.4",
     "console-log-json": "github:SAPikachu/console-log-json",
     "core-js": "^3.22",
-    "dotenv": "^16.0.0",
+    "dotenv": "^16.0.3",
     "elliptic": "^6.5.4",
     "ethers": "^5.7.0",
     "ethers-eip712": "^0.2.0",
     "exponential-backoff": "^3.1.0",
-    "grindery-nexus-common-utils": "^1.0.47",
+    "formidable": "^3.2.5",
+    "fs": "^0.0.1-security",
+    "grindery-nexus-common-utils": "file:../grindery-nexus-common-utils",
     "jose": "^4.10.0",
     "lodash": "^4.17.21",
     "mutexify": "^1.4.0",
     "near-api-js": "^0.45.1",
+    "open": "^8.4.0",
+    "superagent": "^8.0.3",
+    "util": "^0.12.5",
     "uuid": "^8.3.2",
-    "web3": "^1.7.4"
+    "web3": "^1.7.4",
+    "ws": "^8.10.0"
   },
   "devDependencies": {
     "@onflow/types": "^1.0.2",
@@ -40,9 +48,10 @@
     "@types/lodash": "^4.14.182",
     "@types/mutexify": "^1.2.0",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.19.0",
-    "@typescript-eslint/parser": "^5.19.0",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
+    "@typescript-eslint/parser": "^5.41.0",
     "eslint": "^8.13",
+    "eslint-plugin-react": "^7.31.10",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.3"
   }
