+++ src/web3/algorand/clear.teal
@@ -1,0 +1,3 @@
+#pragma version 6
+int 0
+return
