+++ src/web3/evm/call.ts
@@ -1,15 +1,17 @@
 import { ConnectorInput, ConnectorOutput } from "grindery-nexus-common-utils/dist/connector";
 import { TransactionConfig } from "web3-core";
-import { getUserAddress, parseFunctionDeclaration, HUB_ADDRESS } from "./utils";
+import { parseFunctionDeclaration } from "./utils";
 import { getWeb3 } from "./web3";
 import { encodeExecTransaction, execTransactionAbi } from "./gnosisSafe";
-import { parseUserAccessToken } from "../../jwt";
+import { hmac, parseUserAccessToken, TAccessToken } from "../../jwt";
 import axios, { AxiosResponse } from "axios";
 import Web3 from "web3";
 import mutexify from "mutexify/promise";
 
 import GrinderyNexusDrone from "./abi/GrinderyNexusDrone.json";
 import GrinderyNexusHub from "./abi/GrinderyNexusHub.json";
+
+const HUB_ADDRESS = "0xC942DFb6cC8Aade0F54e57fe1eD4320411625F8B";
 
 const hubAvailability = new Map<string, boolean>();
 
@@ -79,6 +81,27 @@ async function prepareRoutedTransaction<T extends Partial<TransactionConfig> | T
   return { tx, droneAddress };
 }
 
+async function getUserAddress(user: TAccessToken, web3: Web3) {
+  let userAddress: string;
+  if ("workspace" in user) {
+    userAddress = web3.utils.toChecksumAddress(
+      "0x" + (await hmac("grindery-web3-address-workspace/" + user.workspace)).subarray(0, 20).toString("hex")
+    );
+  } else {
+    const addressMatch = /^eip155:\d+:(0x.+)$/.exec(user.sub || "");
+    if (addressMatch) {
+      userAddress = addressMatch[1];
+      if (!web3.utils.isAddress(userAddress)) {
+        throw new Error("Unexpected eip155 user ID format");
+      }
+    } else {
+      userAddress = web3.utils.toChecksumAddress(
+        "0x" + (await hmac("grindery-web3-address-sub/" + user.sub)).subarray(0, 20).toString("hex")
+      );
+    }
+  }
+  return userAddress;
+}
 export async function callSmartContract(
   input: ConnectorInput<{
     chain: string;
@@ -98,7 +121,7 @@ export async function callSmartContract(
   }
   const { web3, close, ethersProvider } = getWeb3(input.fields.chain);
   try {
-    const userAddress = await getUserAddress(user);
+    const userAddress = await getUserAddress(user, web3);
     web3.eth.transactionConfirmationBlocks = 1;
     if (!web3.defaultAccount) {
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
@@ -146,7 +169,7 @@ export async function callSmartContract(
       data: callData,
     };
     const isSimulation = functionInfo.constant || functionInfo.stateMutability === "pure" || input.fields.dryRun;
-    const feeData = await ethersProvider.getFeeData();
+    const block = await web3.eth.getBlock("pending").catch(() => web3.eth.getBlock("latest"));
     if (!transactionMutexes[input.fields.chain]) {
       transactionMutexes[input.fields.chain] = safeMutexify();
     }
@@ -205,18 +228,13 @@ export async function callSmartContract(
       const gas = await web3.eth.estimateGas(txConfig);
       txConfig.gas = Math.ceil(gas * 1.1 + 10000);
       let minFee: number;
-      if (input.fields.chain === "eip155:42161") {
-        // Arbitrum, fixed fee
-        txConfig.maxPriorityFeePerGas = 0;
-        txConfig.maxFeePerGas = Number(web3.utils.toWei("110", "kwei"));
-        minFee = txConfig.maxFeePerGas;
-      } else if (feeData.maxFeePerGas && feeData.maxPriorityFeePerGas) {
-        minFee = (feeData.lastBaseFeePerGas || feeData.maxFeePerGas.div(2)).mul(15).div(10).toNumber();
-        const maxTip =
-          input.fields.maxPriorityFeePerGas || Math.floor(minFee / 2);
+      if (block.baseFeePerGas) {
+        const baseFee = Number(block.baseFeePerGas);
+        minFee = baseFee + Number(web3.utils.toWei("30", "gwei"));
+        const maxTip = input.fields.maxPriorityFeePerGas || web3.utils.toWei("75", "gwei");
         const maxFee = input.fields.gasLimit
           ? Math.floor(Number(input.fields.gasLimit) / txConfig.gas)
-          : feeData.maxFeePerGas.add(maxTip).toNumber();
+          : baseFee + Number(maxTip);
         if (maxFee < minFee) {
           throw new Error(
             `Gas limit of ${web3.utils.fromWei(
@@ -226,9 +244,9 @@ export async function callSmartContract(
           );
         }
         txConfig.maxFeePerGas = maxFee;
-        txConfig.maxPriorityFeePerGas = Number(maxTip);
+        txConfig.maxPriorityFeePerGas = Math.min(Number(maxTip), maxFee - baseFee - 1);
       } else {
-        const gasPrice = (feeData.gasPrice || await ethersProvider.getGasPrice()).mul(12).div(10);
+        const gasPrice = await ethersProvider.getGasPrice();
         txConfig.gasPrice = gasPrice.toString();
         minFee = gasPrice.mul(txConfig.gas).toNumber();
       }
