+++ src/web3/flow/index.ts
@@ -3,8 +3,6 @@ import axios from "axios";
 import { ConnectorInput, ConnectorOutput, TriggerBase } from "grindery-nexus-common-utils/dist/connector";
 import { InvalidParamsError } from "grindery-nexus-common-utils/dist/jsonrpc";
 import { sendTransaction } from "./send";
-import blockingTracer from "../../blockingTracer";
-import { TAccessToken } from "../../jwt";
 
 type Block = {
   header: {
@@ -64,7 +62,6 @@ class EventAggregator {
   constructor(private startHeight: number, private contractAddress: string) {}
 
   async fetchEvent(type: string) {
-    blockingTracer.tag("flow.EventAggregator.fetchEvent");
     if (this.fetchPromises.has(type)) {
       return this.fetchPromises.get(type);
     }
@@ -81,7 +78,6 @@ class EventAggregator {
     }
   }
   async fetchEventImpl(type: string) {
-    blockingTracer.tag("flow.EventAggregator.fetchEventImpl");
     if (this.addedEvents.has(type)) {
       return;
     }
@@ -130,7 +126,6 @@ class EventAggregator {
     }
   }
   async forEach(type: string, callback: (events: { [key: string]: ParsedEvent[] }) => Promise<unknown>) {
-    blockingTracer.tag("flow.EventAggregator.forEach");
     await this.fetchEvent(type);
     for (const transaction of this.transactions.values()) {
       if (transaction[type]) {
@@ -149,7 +144,6 @@ class ContractSubscriber extends EventEmitter {
     this.setMaxListeners(1000);
   }
   async main() {
-    blockingTracer.tag("flow.ContractSubscriber.main");
     if (this.running) {
       return;
     }
@@ -206,7 +200,6 @@ class ContractSubscriber extends EventEmitter {
       };
     });
     const handler = async (aggregator: EventAggregator) => {
-      blockingTracer.tag("flow.ContractSubscriber.subscribe.handler");
       await aggregator.forEach(primaryEvent, async (events) => {
         for (const event of events[primaryEvent] || []) {
           const eventFields = event.body.value.fields;
@@ -438,8 +431,4 @@ export async function callSmartContract(
     },
   });
   return { key: input.key, sessionId: input.sessionId, payload: result };
-}
-
-export async function getUserDroneAddress(_user: TAccessToken): Promise<string> {
-  throw new Error("Not implemented");
 }
